changed temperature type to double
git-svn-id: https://svn.cebitec.uni-bielefeld.de/svn/mgx@261 2f99df6c-41e9-6383-b5e2-de53e728a4e2

+++ MGX-Datamodel/src/main/java/de/cebitec/mgx/gui/datamodel/Sample.java
@@ -11,7 +11,7 @@ public class Sample extends ModelBase {
 
     protected Date collectiondate;
     protected String material;
-    protected int temperature;
+    protected double temperature;
     protected int volume;
     protected String volume_unit;
     protected Collection<DNAExtract> extracts;
@@ -44,11 +44,11 @@ public class Sample extends ModelBase {
         return this;
     }
 
-    public int getTemperature() {
+    public double getTemperature() {
         return temperature;
     }
 
-    public Sample setTemperature(int temperature) {
+    public Sample setTemperature(double temperature) {
         this.temperature = temperature;
         return this;
     }
